requirements.txt
discord.py==1.7.3
python-dotenv>=0.9.1
pytz==2018.5
requests==2.20.0
beautifulsoup4==4.6.3
psycopg2-binary==2.8.4
